+++ app/Http/Livewire/GlobalStatistics.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire;
 
+use App\Models\Volume;
 use Illuminate\Support\Facades\DB;
 use Livewire\Component;
 
@@ -11,33 +12,53 @@ class GlobalStatistics extends Component
 
     public array $articleStatistics = [];
 
+    private string $search;
+
     protected $listeners = [
         '$refresh',
+        'search' => 'filter',
     ];
 
     public function render()
     {
-        $volumeStatistics = DB::table('volumes')->select([
-            DB::raw('sum(case when status = 0 then 1 else 0 end) as new'),
-            DB::raw('sum(case when status = 1 then 1 else 0 end) as ordered'),
-            DB::raw('sum(case when status = 2 then 1 else 0 end) as shipped'),
-            DB::raw('sum(case when status = 3 then 1 else 0 end) as delivered'),
-            DB::raw('sum(case when status = 4 then 1 else 0 end) as `read`'),
-            DB::raw('sum(case when status = 3 OR status = 4 then price else 0 end) as price'),
-            DB::raw('count(*) as total'),
-        ])->first();
-        $articleStatistics = DB::table('articles')->select([
-            DB::raw('sum(case when status = 0 then 1 else 0 end) as new'),
-            DB::raw('sum(case when status = 1 then 1 else 0 end) as ordered'),
-            DB::raw('sum(case when status = 2 then 1 else 0 end) as shipped'),
-            DB::raw('sum(case when status = 3 then 1 else 0 end) as delivered'),
-            DB::raw('sum(case when status = 3 then price else 0 end) as price'),
+        $volumes = DB::table('volumes')
+                   ->join('series', 'volumes.series_id', '=', 'series.id')
+                   ->join('publishers', 'series.publisher_id', '=', 'publishers.id');
+        if (!empty($this->search)) {
+            $volumes->where('isbn', 'like', '%' . $this->search . '%')
+            ->orWhere('series.name', 'like', '%' . $this->search . '%')
+            ->orWhere('publishers.name', 'like', '%' . $this->search . '%');
+        }
+        $volumeStatistics = $volumes->get([
+            DB::raw('COALESCE(sum(case when volumes.status = 0 then 1 else 0 end), 0) as new'),
+            DB::raw('COALESCE(sum(case when volumes.status = 1 then 1 else 0 end), 0) as ordered'),
+            DB::raw('COALESCE(sum(case when volumes.status = 2 then 1 else 0 end), 0) as shipped'),
+            DB::raw('COALESCE(sum(case when volumes.status = 3 then 1 else 0 end), 0) as delivered'),
+            DB::raw('COALESCE(sum(case when volumes.status = 4 then 1 else 0 end), 0) as `read`'),
+            DB::raw('COALESCE(sum(case when volumes.status = 3 OR volumes.status = 4 then price else 0 end), 0) as price'),
             DB::raw('count(*) as total'),
         ])->first();
 
+        $articles = DB::table('articles');
+        if (!empty($this->search)) {
+            $articles->where('name', 'like', '%' . $this->search . '%');
+        }
+        $articleStatistics = $articles->select([
+            DB::raw('COALESCE(sum(case when status = 0 then 1 else 0 end), 0) as new'),
+            DB::raw('COALESCE(sum(case when status = 1 then 1 else 0 end), 0) as ordered'),
+            DB::raw('COALESCE(sum(case when status = 2 then 1 else 0 end), 0) as shipped'),
+            DB::raw('COALESCE(sum(case when status = 3 then 1 else 0 end), 0) as delivered'),
+            DB::raw('COALESCE(sum(case when status = 3 then price else 0 end), 0) as price'),
+            DB::raw('count(*) as total'),
+        ])->first();
         $this->volumeStatistics = json_decode(json_encode($volumeStatistics), true);
         $this->articleStatistics = json_decode(json_encode($articleStatistics), true);
 
         return view('livewire.global-statistics');
+    }
+
+    public function filter($filter): void
+    {
+        $this->search = $filter;
     }
 }