+++ app/Http/Livewire/Series/UpcomingSeries.php
@@ -16,11 +16,12 @@ class UpcomingSeries extends Component
 
     public function render()
     {
-        $upcoming = Volume::with('series')->where('ignore_in_upcoming', 'false')->whereIn('status', [0, 1, 2])->orderBy('publish_date')->get();
+        $upcoming = Volume::with('series.publisher')->where('ignore_in_upcoming', 'false')->whereIn('status', [0, 1, 2])->orderBy('publish_date')->get();
         if (!empty($this->search)) {
             $upcoming = $upcoming->filter(function ($volume) {
                 return Str::contains(Str::lower($volume->name), Str::lower($this->search))
-                || Str::contains(Str::lower($volume->isbn), Str::lower($this->search));
+                || Str::contains(Str::lower($volume->isbn), Str::lower($this->search))
+                || Str::contains(Str::lower($volume->series->publisher?->name), Str::lower($this->search));
             });
         }
         $this->upcoming = $upcoming;